+++ src/game/server/entities/light.cpp
@@ -67,7 +67,7 @@ void CLight::Step()
 {
 	Move();
 	vec2 dir(sin(m_Rotation), cos(m_Rotation));
-	vec2 to2 = m_Pos + normalize(dir) * m_CurveLength;
+	vec2 to2 = m_Pos + normalize(dir) * (float)m_CurveLength;
 	GameServer()->Collision()->IntersectNoLaser(m_Pos, to2, &m_To, 0);
 }
